Move primitive types to Objects, reaarange some elements so that NonNulls go first

+++ house-seeker-domain/src/main/java/pl/kala/houseseekerdomain/database/model/document/locality/Locality.java
@@ -19,7 +19,7 @@ public class Locality {
 
     private Voivodship voivodship;
 
-    private int distanceFromHome;
+    private Double distanceFromHome;
 
-    private int distanceFromRailStation;
+    private Double distanceFromRailStation;
 }
